- Implement create movie

+++ Movies.Contracts/IMovieGrainClient.cs
@@ -5,7 +5,7 @@ namespace Movies.Contracts
 {
 	public interface IMovieGrainClient
 	{
-		Task<MovieModel> Get(int id);
+		Task<MovieModel> Get(long id);
 		Task Set(int id, string name, string description, string img, string key, string length, decimal rate);
 	}
 }
